[boost] Bump to 1.68 for chrono fixes and drop patches

+++ include/airmap/date_time.h
@@ -86,11 +86,11 @@ class AIRMAP_EXPORT DateTime {
   friend std::string boost_iso::to_iso_string(const DateTime &datetime);
 };
 
-AIRMAP_EXPORT Hours hours(std::int32_t raw);
-AIRMAP_EXPORT Minutes minutes(std::int32_t raw);
-AIRMAP_EXPORT Seconds seconds(std::int32_t raw);
-AIRMAP_EXPORT Milliseconds milliseconds(std::int32_t raw);
-AIRMAP_EXPORT Microseconds microseconds(std::int32_t raw);
+AIRMAP_EXPORT Hours hours(std::int64_t raw);
+AIRMAP_EXPORT Minutes minutes(std::int64_t raw);
+AIRMAP_EXPORT Seconds seconds(std::int64_t raw);
+AIRMAP_EXPORT Milliseconds milliseconds(std::int64_t raw);
+AIRMAP_EXPORT Microseconds microseconds(std::int64_t raw);
 
 namespace detail {
 
@@ -115,11 +115,11 @@ class AIRMAP_EXPORT Duration {
   friend Microseconds DateTime::operator-(const DateTime &) const;
   friend Microseconds DateTime::time_of_day() const;
 
-  friend Hours airmap::hours(std::int32_t raw);
-  friend Minutes airmap::minutes(std::int32_t raw);
-  friend Seconds airmap::seconds(std::int32_t raw);
-  friend Milliseconds airmap::milliseconds(std::int32_t raw);
-  friend Microseconds airmap::microseconds(std::int32_t raw);
+  friend Hours airmap::hours(std::int64_t raw);
+  friend Minutes airmap::minutes(std::int64_t raw);
+  friend Seconds airmap::seconds(std::int64_t raw);
+  friend Milliseconds airmap::milliseconds(std::int64_t raw);
+  friend Microseconds airmap::microseconds(std::int64_t raw);
 };
 
 }  // namespace detail
